the_app/sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomi-sdk"
version = "0.1.0"
description = "Quality-first multi-agent framework with built-in safety"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Autonomi", email = "sdk@autonomi.dev" }
]
keywords = [
    "ai",
    "agents",
    "multi-agent",
    "llm",
    "anthropic",
    "openai",
    "autonomous"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.32.0",
    "openai>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
memory = [
    "aiosqlite>=0.19.0",
    "asyncpg>=0.28.0",
]
telemetry = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
all = [
    "autonomi-sdk[dev,memory,telemetry]",
]

[project.urls]
Homepage = "https://github.com/autonomi/autonomi-sdk"
Documentation = "https://docs.autonomi.dev/sdk"
Repository = "https://github.com/autonomi/autonomi-sdk"

[tool.hatch.build.targets.wheel]
packages = ["src/autonomi"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
